refactor(search): move the Audible series endpoint into its own partial
ActiveProductionSourceFiles_RemainFocused caps production files at 500
lines and SearchController.cs reached 506 with the typed series endpoint
inlined. Extracted to a partial alongside the existing
SearchController.DownloadReferences.cs, which brings the controller back
to 457 lines.
